newsletter-bot/requirements.txt
python-dotenv
yfinance
pandas
numpy
scipy
vaderSentiment
matplotlib
fpdf
reportlab  # Optional for advanced PDF
scikit-learn
groq